+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `OxidEsales\EshopCommunity\Core\ViewConfig::getShowSuggest()`
 - New settings:
   - `blAllowSuggestArticle` - now it's possible to disable recommendation feature in admin.
+- New methods:
+  - `OxidEsales\EshopCommunity\Application\Model\User::isMallAdmin()`
 
 ### Changed